+++ packages/snippets/src/browser/translateTo.js
@@ -2,8 +2,8 @@ const setElementStyleProperties = require('./setElementStyleProperties')
 
 function translateTo([element, offset] = []) {
   element = element || document.documentElement
-  const value = `translate(${-offset.x}px, ${-offset.y}px)`
-  setElementStyleProperties([element, {transform: {value}, '-webkit-transform': {value}}])
+  const transform = {value: `translate(${-offset.x}px, ${-offset.y}px)`, important: true}
+  setElementStyleProperties([element, {transform, '-webkit-transform': transform}])
   return offset
 }