+++ packages/spec-driver-webdriverio/src/spec-driver.ts
@@ -210,7 +210,7 @@ export async function getCapabilities(browser: Driver): Promise<Record<string, a
   return browser.getSession?.() ?? browser.capabilities
 }
 export async function getDriverInfo(browser: Driver): Promise<DriverInfo> {
-  return {sessionId: browser.sessionId}
+  return {sessionId: browser.sessionId, isNative: !/^WEBVIEW_/.test(await browser.getContext())}
 }
 export async function getTitle(browser: Driver): Promise<string> {
   return browser.getTitle()