+++ packages/snippets/CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ## Unreleased
 
+- overwrite `transform` style property with `!important` in `translateTo`
 
 ## 2.1.9 - 2021/12/17